requirements.txt
customtkinter
Pillow
pynput